changing structure, moving db to src
