+++ src/db/queries/tweets/delete-tweet.js
