+++ src/db/queries/tweets/get-tweet-by-id.js
