+++ src/services/tweets.service.js
@@ -1,7 +1,7 @@
-const getAllTweets = require('../../db/queries/tweets/get-all-tweets');
-const getTweetById = require('../../db/queries/tweets/get-tweet-by-id');
-const createTweet = require('../../db/queries/tweets/create-tweet');
-const deleteTweet = require('../../db/queries/tweets/delete-tweet');
+const getAllTweets = require('../db/queries/tweets/get-all-tweets');
+const getTweetById = require('../db/queries/tweets/get-tweet-by-id');
+const createTweet = require('../db/queries/tweets/create-tweet');
+const deleteTweet = require('../db/queries/tweets/delete-tweet');
 
 module.exports = {
     getAllTweetsService() {
