+++ tests/database/gallery/delete.js
@@ -1,5 +1,5 @@
 const assert = require('assert');
-const Gallery = require('../../../db/models/gallery');
+const Gallery = require('../../../src/db/models/gallery');
 
 describe('Deleting a gallery', () => {
     let gallery;
