+++ tests/database/gallery/validation.js
@@ -1,5 +1,5 @@
 const assert = require('assert');
-const Gallery = require('../../../db/models/gallery');
+const Gallery = require('../../../src/db/models/gallery');
 
 describe('Validating gallery records', () => {
     it('Image URL must be less than 200 characters', (done) => {
