+++ tests/database/tweet/read.js
@@ -1,6 +1,6 @@
 const assert = require('assert');
-const Tweet = require('../../../db/models/tweets');
-const User = require('../../../db/models/user');
+const Tweet = require('../../../src/db/models/tweets');
+const User = require('../../../src/db/models/user');
 
 describe('Reading records', () => {
     let user;
